requirements.txt
pytest
playwright
allure-pytest
pytest-check
